Move app header
