+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import './App.css';
-import AppHeader from './components/app-header';
+import AppHeader from './components/header/app-header';
 import BurgerIngredients from './components/burger-ingredients';
 import BurgerConstructor from './components/burger-constructor';
 import data from './utils/data';
